requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
apturl==0.5.2
argcomplete==1.8.1
async-timeout==4.0.1
attrs==21.2.0
Babel==2.8.0
bcc==0.18.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
cupshelpers==1.0
dbus-fast==2.44.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.3.0
fastapi==0.121.0
fasteners==0.14.1
fido2==0.9.1
frozenlist==1.2.0
future==0.18.2
h11==0.16.0
httplib2==0.20.2
httptools==0.7.1
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
livereload==2.6.3
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Markdown==3.3.6
MarkupSafe==2.0.1
mkdocs==1.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==5.1.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipx==1.0.0
pluginbase==1.0.1
protobuf==3.12.4
proton-core==0.7.3
proton-keyring-linux==0.2.1
proton-vpn-api-core==4.13.1
proton-vpn-daemon==0.13.4
proton-vpn-gtk-app==4.12.0
proton-vpn-lib==0.1.1
proton-vpn-network-manager==0.13.4
psutil==5.9.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
python-gnupg==0.4.8
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
sentry-sdk==1.4.3
six==1.16.0
sniffio==1.3.1
starlette==0.49.3
systemd-python==234
tornado==6.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
userpath==1.8.0
uvicorn==0.38.0
uvloop==0.22.1
vboxapi==1.0
wadllib==1.3.6
wafw00f==2.1.0
watchfiles==1.1.1
websockets==15.0.1
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
